exchange_rate/pyproject.toml
[tool.poetry]
name = "google_api"
version = "0.1.0"
description = ""
authors = ["maximq"]

[tool.poetry.dependencies]
python = "^3.10"
google = "^3.0.0"
google-api-python-client = "^2.62.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^0.5.3"
python-dotenv = "^0.21.0"
gspread = "^5.5.0"
oauth2client = "^4.1.3"
Django = "^4.1.1"
djangorestframework = "^3.14.0"
gunicorn = "^20.1.0"
psycopg2-binary = "^2.9.3"
django-environ = "^0.9.0"
django-filter = "^22.1"
celery = "^5.2.7"
redis = "^4.3.4"
django-redis = "^5.2.0"
aiohttp = "^3.8.3"
asyncio = "^3.4.3"
django-cors-headers = "^3.13.0"
python-telegram-bot = "^13.14"
djangochannelsrestframework = "1.0.0"
channels = "^3.0.5"
channels-redis = "^3.4.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
